Refactor organization and cleanups
