+++ apps/web/app/dashboard/caps/components/SelectedCapsBar.tsx
@@ -1,0 +1,73 @@
+"use client";
+
+import { Button } from "@cap/ui";
+import { faTrash } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { AnimatePresence, motion } from "framer-motion";
+import NumberFlow from "@number-flow/react";
+
+interface SelectedCapsBarProps {
+  selectedCaps: string[];
+  setSelectedCaps: (caps: string[]) => void;
+  deleteSelectedCaps: () => void;
+  isDeleting: boolean;
+}
+
+export const SelectedCapsBar = ({
+  selectedCaps,
+  setSelectedCaps,
+  deleteSelectedCaps,
+  isDeleting
+}: SelectedCapsBarProps) => {
+  return (
+    <AnimatePresence>
+      {selectedCaps.length > 0 && (
+        <motion.div
+          className="flex fixed right-0 left-0 bottom-4 z-50 justify-between items-center px-6 py-3 mx-auto w-full max-w-xl rounded-xl border shadow-lg border-gray-3 bg-gray-2"
+          initial={{ opacity: 0, y: 10, scale: 0.9 }}
+          animate={{ opacity: 1, y: 0, scale: 1 }}
+          exit={{
+            opacity: 0,
+            y: 10,
+            scale: 0.9,
+            transition: { duration: 0.2 },
+          }}
+          transition={{
+            type: "spring",
+            damping: 15,
+            stiffness: 200,
+          }}
+        >
+          <div className="flex gap-1 text-sm font-medium text-gray-10">
+            <NumberFlow
+              value={selectedCaps.length}
+              className="tabular-nums text-md text-gray-12"
+            />
+            cap{selectedCaps.length !== 1 ? "s" : ""} selected
+          </div>
+          <div className="flex gap-2 ml-4">
+            <Button
+              variant="dark"
+              onClick={() => setSelectedCaps([])}
+              className="text-sm"
+              size="sm"
+            >
+              Cancel
+            </Button>
+            <Button
+              style={{ minWidth: "auto" }}
+              variant="destructive"
+              onClick={deleteSelectedCaps}
+              disabled={isDeleting}
+              className="text-sm w-[40px]"
+              spinner={isDeleting}
+              size="sm"
+            >
+              <FontAwesomeIcon icon={faTrash} />
+            </Button>
+          </div>
+        </motion.div>
+      )}
+    </AnimatePresence>
+  );
+};
