+++ apps/web/app/dashboard/shared-caps/SharedCaps.tsx
@@ -22,11 +22,9 @@ type SharedVideoData = {
 export const SharedCaps = ({
   data,
   count,
-  activeOrganizationId,
 }: {
   data: SharedVideoData;
   count: number;
-  activeOrganizationId: string;
 }) => {
   const params = useSearchParams();
   const page = Number(params.get("page")) || 1;
